% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits_extractors.R
\name{extract_intercepts}
\alias{extract_intercepts}
\title{Extract Intercepts from Bootstrap-Replicated Models}
\usage{
extract_intercepts(fits)
}
\arguments{
\item{fits}{A list of fitted \code{cv.glmnet} objects.}
}
\value{
A numeric vector of intercepts (one per bootstrap).
}
\description{
Extract Intercepts from Bootstrap-Replicated Models
}
