% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits_extractors.R
\name{extract_lambda_min}
\alias{extract_lambda_min}
\title{Extract Optimal Lambda Values from Bootstrap Models}
\usage{
extract_lambda_min(fits)
}
\arguments{
\item{fits}{A list of fitted \code{cv.glmnet} objects.}
}
\value{
A numeric vector of optimal lambda values (i.e., \code{lambda.min} from each fit).
}
\description{
Extract Optimal Lambda Values from Bootstrap Models
}
