% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits_extractors.R
\name{selection_frequency}
\alias{selection_frequency}
\title{Compute Variable Selection Frequencies Across Bootstraps}
\usage{
selection_frequency(fits)
}
\arguments{
\item{fits}{A list of fitted \code{cv.glmnet} objects.}
}
\value{
A named numeric vector with selection frequencies (values between 0 and 1).
}
\description{
Calculates how often each coefficient is non-zero across a list or matrix of
coefficient vectors.
}
